requirements.txt
numpy~=1.26.4
tensorflow~=2.15.0
tqdm~=4.64.0
streamlit~=1.24.0
pandas~=1.4.2
Pillow~=9.0.1
scikit-learn~=1.0.2
opencv-python~=4.9.0.80